feat: interfaces-b10-cli-generator-006 - Verify gates and merge prerequisite
